Revert "Fix: crash if value of bound GSettings key is not in enum"
This reverts commit 9777a1f5b28c6995791994c4abe91902533a1939.

The original commit was trying to fix issue #118 but from the comment
<https://github.com/realmazharhussain/gdm-settings/issues/118#issuecomment-1264380025>,
it seems like it does not do its job because of the problem being with
the app not being completely relocatable.

Also, the original commit would have caused a lot of pain in the future
where Enum members named wrongly by mistake would have been silently ignored
and caused misbehavior instead of an Exception being thrown.